requirements.txt
pillow
cryptography
segno
opencv-python
pyzbar
numpy
PySide6

